# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para envio de mensagens via WhatsApp integrada ao AgroServer",
        "title": "AgroChat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte AgroChat",
            "email": "suporte@agrochat.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/enviar-mensagem": {
            "post": {
                "description": "Envia uma mensagem de texto simples para um número de WhatsApp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mensagens"
                ],
                "summary": "Enviar mensagem simples",
                "parameters": [
                    {
                        "description": "Número e mensagem",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "mensagem": {
                                    "type": "string"
                                },
                                "numero": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagem enviada com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro no servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "WhatsApp não conectado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/enviar-verificacao": {
            "post": {
                "description": "Envia código de verificação formatado para clientes do AgroServer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verificação"
                ],
                "summary": "Enviar código de verificação",
                "parameters": [
                    {
                        "description": "Dados da verificação",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.VerificacaoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verificação enviada com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro no servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "WhatsApp não conectado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/whatsapp/send": {
            "post": {
                "description": "Envia uma mensagem para um número de WhatsApp com verificações detalhadas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp"
                ],
                "summary": "Envia mensagem via WhatsApp (endpoint antigo)",
                "parameters": [
                    {
                        "description": "Dados da mensagem",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "phone": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagem enviada com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro no servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "WhatsApp não conectado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/whatsapp/status": {
            "get": {
                "description": "Retorna informações sobre o status da conexão com o WhatsApp",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp"
                ],
                "summary": "Status da conexão WhatsApp",
                "responses": {
                    "200": {
                        "description": "Status da conexão",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.VerificacaoRequest": {
            "type": "object",
            "required": [
                "codigoVerificacao",
                "documento",
                "metodo",
                "nomeCliente",
                "nomeVendedor",
                "telefone"
            ],
            "properties": {
                "codigoVerificacao": {
                    "type": "string"
                },
                "documento": {
                    "type": "string"
                },
                "endereco": {
                    "type": "string"
                },
                "metodo": {
                    "type": "string"
                },
                "nomeCliente": {
                    "type": "string"
                },
                "nomeVendedor": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                }
            }
        }
    }
}